+++ CHANGELOG.md
@@ -1,6 +1,8 @@
 # Release Notes
 
 ## v2.7.0 (2024-04-23)
+- Deleted the old random videos from the home page sidebar
+- Added 5 new random videos to the home page sidebar
 
 ## v2.6.1 (2024-04-21)
 - Moved images from the external server to the extension's folder. Now all images are served from the extension's rather than the `https://php-revival.github.io`. It doesn't effect anything, it's just a better practice